[#96] chore : 하이드레이션 오류 수정
